lib: refactor floors
